Added assertions for the cookie banner on AMP
