+++ src/app/containers/ConsentBanner/a11y.test.jsx
@@ -1,0 +1,141 @@
+import React, { createRef } from 'react';
+import { render } from '@testing-library/react';
+import ConsentBanner from './index';
+import { UserContextProvider } from '#contexts/UserContext';
+import { ToggleContext } from '#contexts/ToggleContext';
+import { ServiceContext } from '#contexts/ServiceContext';
+import { service as pidginServiceConfig } from '#lib/config/services/pidgin';
+import { RequestContextProvider } from '#contexts/RequestContext';
+import { FRONT_PAGE } from '#app/routes/utils/pageTypes';
+
+const defaultToggleState = {
+  chartbeatAnalytics: {
+    enabled: false,
+  },
+};
+const mockToggleDispatch = jest.fn();
+
+// eslint-disable-next-line react/prop-types
+const AmpBannerWithContext = ({ service, serviceConfig, variant }) => {
+  return (
+    <RequestContextProvider
+      isAmp
+      pageType={FRONT_PAGE}
+      pathname="/"
+      service={service}
+    >
+      <ToggleContext.Provider
+        value={{
+          toggleState: defaultToggleState,
+          toggleDispatch: mockToggleDispatch,
+        }}
+      >
+        <UserContextProvider>
+          <ServiceContext.Provider value={serviceConfig[variant]}>
+            <ConsentBanner />
+          </ServiceContext.Provider>
+        </UserContextProvider>
+      </ToggleContext.Provider>
+    </RequestContextProvider>
+  );
+};
+
+const CanonicalBannerWithContext = React.forwardRef(
+  // eslint-disable-next-line react/prop-types
+  ({ serviceConfig, variant }, ref) => {
+    return (
+      <>
+        <div ref={ref}>
+          <a href="/">BBC Brand</a>
+        </div>
+        <ToggleContext.Provider
+          value={{
+            toggleState: defaultToggleState,
+            toggleDispatch: mockToggleDispatch,
+          }}
+        >
+          <UserContextProvider>
+            <ServiceContext.Provider value={serviceConfig[variant]}>
+              <ConsentBanner onDismissFocusRef={ref} />
+            </ServiceContext.Provider>
+          </UserContextProvider>
+        </ToggleContext.Provider>
+      </>
+    );
+  },
+);
+
+describe('canonical', () => {
+  it('should focus on canonical consent banner heading on mount on canonical', () => {
+    const { getByText } = render(
+      <CanonicalBannerWithContext
+        serviceConfig={pidginServiceConfig}
+        variant="default"
+      />,
+    );
+    const pidginPrivacyHeading =
+      pidginServiceConfig.default.translations.consentBanner.privacy.title;
+
+    expect(document.activeElement).toBe(getByText(pidginPrivacyHeading));
+  });
+
+  it('should focus on the link within the referenced element after cookie accept on canonical', () => {
+    const onDismissFocusRef = createRef(null);
+    const { getByText } = render(
+      <CanonicalBannerWithContext
+        serviceConfig={pidginServiceConfig}
+        variant="default"
+        ref={onDismissFocusRef}
+      />,
+    );
+
+    const pidginPrivacyAccept =
+      pidginServiceConfig.default.translations.consentBanner.privacy.accept;
+    const pidginCookieAccept =
+      pidginServiceConfig.default.translations.consentBanner.cookie.canonical
+        .accept;
+
+    getByText(pidginPrivacyAccept).click();
+    getByText(pidginCookieAccept).click();
+
+    expect(document.activeElement).toBe(getByText('BBC Brand'));
+  });
+});
+
+describe('amp', () => {
+  it('should render a focussable manage cookies heading on AMP', () => {
+    const { container } = render(
+      <AmpBannerWithContext
+        service="pidgin"
+        serviceConfig={pidginServiceConfig}
+        variant="default"
+      />,
+    );
+
+    const manageCookiesHeading = container.querySelector(
+      '#manageCookiesHeading',
+    );
+    manageCookiesHeading.focus();
+
+    expect(document.activeElement).toBe(manageCookiesHeading);
+  });
+
+  it('should render a focussable cookie banner heading on AMP', () => {
+    const { getByText } = render(
+      <AmpBannerWithContext
+        service="pidgin"
+        serviceConfig={pidginServiceConfig}
+        variant="default"
+      />,
+    );
+
+    const pidginCookieAcceptAmp =
+      pidginServiceConfig.default.translations.consentBanner.cookie.amp.initial
+        .title;
+
+    const pidginCookieHeading = getByText(pidginCookieAcceptAmp);
+    pidginCookieHeading.focus();
+
+    expect(document.activeElement).toBe(pidginCookieHeading);
+  });
+});
